Add token argument and deprecate checkoutId argument in CheckoutLinesAdd and CheckoutLinesUpdate mutations
